+++ record_android/android/src/main/kotlin/com/llfbandit/record/record/encoder/MediaCodecEncoder.kt
@@ -16,19 +16,28 @@ class MediaCodecEncoder(
 
     private val codec = createCodec(mediaFormat)
     private var trackIndex = -1
-    private var recordStopped: Boolean = false
+    private var recordStopped = false
+    private var recordPaused = false
 
     override fun start() {
         codec.setCallback(this)
         codec.start()
     }
 
+    override fun pause() {
+        recordPaused = true
+    }
+
+    override fun resume() {
+        recordPaused = false
+    }
+
     override fun stop() {
+        recordPaused = false
         recordStopped = true
     }
 
     override fun release() {
-        codec.release()
         container.release()
     }
 
@@ -75,6 +84,8 @@ class MediaCodecEncoder(
 
     private fun internalStop() {
         codec.stop()
+        codec.release()
+        container.stop()
 
         listener.onEncoderStop()
     }
@@ -93,6 +104,11 @@ class MediaCodecEncoder(
     }
 
     override fun onInputBufferAvailable(codec: MediaCodec, index: Int) {
+        if (recordPaused) {
+            codec.queueInputBuffer(index, 0, 0, 0, 0)
+            return
+        }
+
         try {
             val byteBuffer = codec.getInputBuffer(index) ?: return
             val resultBytes = listener.onEncoderDataNeeded(byteBuffer)